+++ Source/BugsnagSessionTrackingPayload.m
@@ -26,6 +26,11 @@
 + (BugsnagClient *)client;
 @end
 
+@interface BugsnagDevice ()
++ (BugsnagDevice *)deviceWithDictionary:(NSDictionary *)event;
+- (NSDictionary *)toDict;
+@end
+
 @interface BugsnagApp ()
 + (BugsnagApp *)appWithDictionary:(NSDictionary *)event
                            config:(BugsnagConfiguration *)config;
@@ -66,7 +71,9 @@
     NSDictionary *event = [self appInfo:systemInfo config:self.config];
     BugsnagApp *app = [BugsnagApp appWithDictionary:event config:self.config];
     BSGDictSetSafeObject(dict, [app toDict], @"app");
-    BSGDictSetSafeObject(dict, BSGParseDeviceState(systemInfo), @"device");
+
+    BugsnagDevice *device = [BugsnagDevice deviceWithDictionary:@{@"system": systemInfo}];
+    BSGDictSetSafeObject(dict, [device toDict], @"device");
     return dict;
 }