+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Bugsnag Notifiers on other platforms.
 
 ## Enhancements
 
+* Convert `event.device` from `NSDictionary` to a structured class
+  [#526](https://github.com/bugsnag/bugsnag-cocoa/pull/526)
+
 * Convert `event.app` from `NSDictionary` to a structured class
   [#520](https://github.com/bugsnag/bugsnag-cocoa/pull/520)